# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface KalshiMarket {
  ticker: string;
  title: string;
  category: string;
  closeTime: string;
  volume: number;
  yesPrice: number;
  noPrice: number;
  liquidity: string;
  status: string;
}

export type KalshiSignalSide = typeof KalshiSignalSide[keyof typeof KalshiSignalSide];


export const KalshiSignalSide = {
  YES: 'YES',
  NO: 'NO',
} as const;

export type KalshiSignalStatus = typeof KalshiSignalStatus[keyof typeof KalshiSignalStatus];


export const KalshiSignalStatus = {
  watch: 'watch',
  strong: 'strong',
  caution: 'caution',
} as const;

export type KalshiSignalRecommendation = typeof KalshiSignalRecommendation[keyof typeof KalshiSignalRecommendation];


export const KalshiSignalRecommendation = {
  favorable: 'favorable',
  wait: 'wait',
  avoid: 'avoid',
} as const;

export interface KalshiSignal {
  id: string;
  ticker: string;
  asset?: string;
  title: string;
  category: string;
  side: KalshiSignalSide;
  entryPrice: number;
  modelProbability: number;
  marketProbability: number;
  edge: number;
  confidence: number;
  liquidity: string;
  timeToClose: string;
  status: KalshiSignalStatus;
  /** Gross expected return before fees. */
  expectedValue: number;
  /** Estimated taker fee per contract using Kalshi's general quadratic fee formula. */
  estimatedFee: number;
  /** Expected return after the estimated taker fee. */
  netExpectedValue: number;
  recommendation: KalshiSignalRecommendation;
  score: number;
  explanation: string;
  updatedAt: string;
}

export interface KalshiPricePoint {
  time: string;
  yesPrice: number;
  volume: number;
}

export interface KalshiPerformancePoint {
  label: string;
  cumulativeReturn: number;
  hitRate: number;
  trades: number;
}

export type KalshiAssetStatsCurrentSide = typeof KalshiAssetStatsCurrentSide[keyof typeof KalshiAssetStatsCurrentSide];


export const KalshiAssetStatsCurrentSide = {
  YES: 'YES',
  NO: 'NO',
} as const;

export type KalshiAssetStatsRecommendation = typeof KalshiAssetStatsRecommendation[keyof typeof KalshiAssetStatsRecommendation];


export const KalshiAssetStatsRecommendation = {
  favorable: 'favorable',
  wait: 'wait',
  avoid: 'avoid',
} as const;

export interface KalshiAssetStats {
  asset: string;
  seriesTicker: string;
  sampleSize: number;
  hitRate: number;
  grossReturn: number;
  netReturn: number;
  averageNetProfitCents: number;
  maxDrawdown: number;
  profitable: boolean;
  currentSide?: KalshiAssetStatsCurrentSide;
  currentProbability?: number;
  currentNetExpectedValue?: number;
  recommendation?: KalshiAssetStatsRecommendation;
  forwardTracked: number;
  forwardResolved: number;
  forwardHitRate: number;
  forwardNetReturn: number;
}

export interface KalshiForwardAssetStats {
  asset: string;
  tracked: number;
  pending: number;
  resolved: number;
  hitRate: number;
  netReturn: number;
  totalProfit: number;
}

export interface KalshiForwardTracking {
  startedAt: string;
  tracked: number;
  pending: number;
  resolved: number;
  hitRate: number;
  netReturn: number;
  totalProfit: number;
  assets: KalshiForwardAssetStats[];
}

export interface KalshiDashboard {
  asOf: string;
  source: string;
  isLive: boolean;
  totalMarkets: number;
  activeSignals: number;
  averageConfidence: number;
  backtestHitRate: number;
  /** Gross paper return in the holdout simulation before fees. */
  simulatedReturn: number;
  /** Estimated paper return after Kalshi's general taker fee formula. */
  netSimulatedReturn: number;
  modelName: string;
  modelTrainedAt: string;
  trainingMarkets: number;
  backtestSampleSize: number;
  /** Accuracy from choosing the side above 50 percent using only the market quote at the same cutoff. */
  baselineHitRate: number;
  /** Mean squared error of the predicted probabilities on the chronological holdout set; lower is better. */
  brierScore: number;
  assetStats: KalshiAssetStats[];
  liveTracking: KalshiForwardTracking;
  signals: KalshiSignal[];
  priceHistory: KalshiPricePoint[];
  performance: KalshiPerformancePoint[];
  markets: KalshiMarket[];
}

export type GetKalshiDashboardParams = {
category?: string;
/**
 * @minimum 0
 * @maximum 100
 */
minConfidence?: number;
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};