# lib/api-zod/src/generated/types/kalshiAssetStats.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KalshiAssetStatsCurrentSide } from './kalshiAssetStatsCurrentSide';
import type { KalshiAssetStatsRecommendation } from './kalshiAssetStatsRecommendation';

export interface KalshiAssetStats {
  asset: string;
  seriesTicker: string;
  sampleSize: number;
  hitRate: number;
  grossReturn: number;
  netReturn: number;
  averageNetProfitCents: number;
  maxDrawdown: number;
  profitable: boolean;
  currentSide?: KalshiAssetStatsCurrentSide;
  currentProbability?: number;
  currentNetExpectedValue?: number;
  recommendation?: KalshiAssetStatsRecommendation;
  forwardTracked: number;
  forwardResolved: number;
  forwardHitRate: number;
  forwardNetReturn: number;
}