# lib/api-zod/src/generated/types/kalshiDashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KalshiAssetStats } from './kalshiAssetStats';
import type { KalshiForwardTracking } from './kalshiForwardTracking';
import type { KalshiMarket } from './kalshiMarket';
import type { KalshiPerformancePoint } from './kalshiPerformancePoint';
import type { KalshiPricePoint } from './kalshiPricePoint';
import type { KalshiSignal } from './kalshiSignal';

export interface KalshiDashboard {
  asOf: string;
  source: string;
  isLive: boolean;
  totalMarkets: number;
  activeSignals: number;
  averageConfidence: number;
  backtestHitRate: number;
  /** Gross paper return in the holdout simulation before fees. */
  simulatedReturn: number;
  /** Estimated paper return after Kalshi's general taker fee formula. */
  netSimulatedReturn: number;
  modelName: string;
  modelTrainedAt: Date;
  trainingMarkets: number;
  backtestSampleSize: number;
  /** Accuracy from choosing the side above 50 percent using only the market quote at the same cutoff. */
  baselineHitRate: number;
  /** Mean squared error of the predicted probabilities on the chronological holdout set; lower is better. */
  brierScore: number;
  assetStats: KalshiAssetStats[];
  liveTracking: KalshiForwardTracking;
  signals: KalshiSignal[];
  priceHistory: KalshiPricePoint[];
  performance: KalshiPerformancePoint[];
  markets: KalshiMarket[];
}