# lib/api-zod/src/generated/types/kalshiForwardTracking.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KalshiForwardAssetStats } from './kalshiForwardAssetStats';

export interface KalshiForwardTracking {
  startedAt: Date;
  tracked: number;
  pending: number;
  resolved: number;
  hitRate: number;
  netReturn: number;
  totalProfit: number;
  assets: KalshiForwardAssetStats[];
}