# lib/api-zod/src/generated/types/kalshiMarket.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KalshiMarket {
  ticker: string;
  title: string;
  category: string;
  closeTime: string;
  volume: number;
  yesPrice: number;
  noPrice: number;
  liquidity: string;
  status: string;
}