# lib/api-zod/src/generated/types/kalshiPerformancePoint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KalshiPerformancePoint {
  label: string;
  cumulativeReturn: number;
  hitRate: number;
  trades: number;
}