# lib/api-zod/src/generated/types/kalshiSignal.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KalshiSignalRecommendation } from './kalshiSignalRecommendation';
import type { KalshiSignalSide } from './kalshiSignalSide';
import type { KalshiSignalStatus } from './kalshiSignalStatus';

export interface KalshiSignal {
  id: string;
  ticker: string;
  asset?: string;
  title: string;
  category: string;
  side: KalshiSignalSide;
  entryPrice: number;
  modelProbability: number;
  marketProbability: number;
  edge: number;
  confidence: number;
  liquidity: string;
  timeToClose: string;
  status: KalshiSignalStatus;
  /** Gross expected return before fees. */
  expectedValue: number;
  /** Estimated taker fee per contract using Kalshi's general quadratic fee formula. */
  estimatedFee: number;
  /** Expected return after the estimated taker fee. */
  netExpectedValue: number;
  recommendation: KalshiSignalRecommendation;
  score: number;
  explanation: string;
  updatedAt: string;
}